Refactor: Move intersection observer to TimeSeriesExplorer
